Struct brotli::bitreader::BitReader

pub struct BitReader<R: Read> {
    inner: BufReader<R>,
    bit_pos: u8,
    current_byte: Option<u8>,
    global_bit_pos: usize,
}

Wrapper for a Reader, providing convenience methods to read the stream bit-by-bit.

Examples

extern crate compression;

use compression::bitreader::BitReader; use std::fs::File;

let f = try!(File::open("filename")); let mut br = BitReader::new(f); let byte: u8 = br.read_u8().unwrap();

Fields

inner

bit_pos

current_byte

global_bit_pos

Methods

impl<R: Read> BitReader<R>

fn new(inner: R) -> BitReader<R>

Creates a BitReader from a Read.

fn read_exact(&mut self, buf: &mut [u8]) -> Result<()>

fn read_u8(&mut self) -> Result<u8, BitReaderError>

Reads a u8 from the stream, reading exactly one byte, or, more precisely, 8 bits, which might not be aligned to byte boundaries. Returns a BitReaderError if the stream ends prematurely.

fn read_u8_from_nibble(&mut self) -> Result<u8, BitReaderError>

Reads a u8 from 4 bits. Returns a BitReaderError if the stream ends prematurely.

fn read_u32_from_n_bits(&mut self, n: usize) -> Result<u32, BitReaderError>

Reads a u32 from n bits. Only supports little endian, i.e. the least significant bit comes first in the stream. Returns a BitReaderError if the stream ends prematurely, or if n exceeds the number of possible bits.

fn read_u32_from_n_nibbles(&mut self, n: usize) -> Result<u32, BitReaderError>

Reads a u32 from n nibbles (4 bits). Only supports little endian, i.e. the least significant nibble comes first in the stream. Returns a BitReaderError if the stream ends prematurely, or if n exceeds the number of possible nibbles.

fn read_bit(&mut self) -> Result<bool, BitReaderError>

Reads one bit from the stream, returns a bool result. Returns a BitReaderError if the stream ends prematurely.

fn read_bit_as_usize(&mut self) -> Result<usize, BitReaderError>

Reads one bit from the stream, returns a usize result. Returns a BitReaderError if the stream ends prematurely.

fn read_u8_from_n_bits(&mut self, n: usize) -> Result<u8, BitReaderError>

Reads a u8 from n bits from the stream. Returns a BitReaderError if the stream ends prematurely, or if n exceeds the possible number of bits.

fn read_u8_from_byte_tail(&mut self) -> Result<u8, BitReaderError>

Reads u8 from bits up to the next byte boundary. Returns a BitReaderError if the stream ends prematurely.

fn read_u16_from_n_bits(&mut self, n: usize) -> Result<u16, BitReaderError>

Reads a u16 from n bits from the stream. Returns a BitReaderError if the stream ends prematurely, or if n exceeds the possible number of bits.

fn read_fixed_length_string(&mut self, len: usize) -> Result<Vec<u8>, BitReaderError>

Reads a vector of u8 of a given length. Returns a BitReaderError if the stream ends prematurely.