Struct brotli::Decompressor

pub struct Decompressor<R: Read> {
    in_stream: BitReader<R>,
    header: Header,
    buf: VecDeque<Literal>,
    output_window: Option<RingBuffer<Literal>>,
    state: State,
    meta_block: MetaBlock,
    count_output: usize,
    literal_buf: RingBuffer<Literal>,
    distance_buf: RingBuffer<Distance>,
}

Wraps an input stream and provides methods for decompressing.

Examples

use std::io::{ Read, stdout, Write };
use brotli::Decompressor;

let brotli_stream = std::fs::File::open("data/64x.compressed").unwrap();

let mut decompressed = &mut Vec::new();
let _ = Decompressor::new(brotli_stream).read_to_end(&mut decompressed);

let mut expected = &mut Vec::new();
let _ = std::fs::File::open("data/64x").unwrap().read_to_end(&mut expected);

assert_eq!(expected, decompressed);

stdout().write_all(decompressed).ok();

Fields

in_stream

header

buf

output_window

state

meta_block

count_output

literal_buf

ring buffer for last 2 literals, gets set at the beginning of the stream, and then lives until the end

distance_buf

ring buffer for last 4 distances, gets set at the beginning of the stream, and then lives until the end

Methods

impl<R: Read> Decompressor<R>

fn new(r: R) -> Decompressor<R>

Creates Decompressor from Read.

fn parse_wbits(&mut self) -> Result<State, DecompressorError>

fn parse_is_last(&mut self) -> Result<State, DecompressorError>

fn parse_is_last_empty(&mut self) -> Result<State, DecompressorError>

fn parse_m_nibbles(&mut self) -> Result<State, DecompressorError>

fn parse_m_skip_bytes(&mut self) -> Result<State, DecompressorError>

fn parse_m_skip_len(&mut self) -> Result<State, DecompressorError>

fn parse_m_len(&mut self) -> Result<State, DecompressorError>

fn parse_is_uncompressed(&mut self) -> Result<State, DecompressorError>

fn parse_mlen_literals(&mut self) -> Result<State, DecompressorError>

fn parse_n_bltypes(&mut self) -> Result<NBltypes, DecompressorError>

fn parse_n_bltypes_l(&mut self) -> Result<State, DecompressorError>

fn parse_n_bltypes_i(&mut self) -> Result<State, DecompressorError>

fn parse_n_bltypes_d(&mut self) -> Result<State, DecompressorError>

fn parse_n_postfix(&mut self) -> Result<State, DecompressorError>

fn parse_n_direct(&mut self) -> Result<State, DecompressorError>

fn parse_context_modes_literals(&mut self) -> Result<State, DecompressorError>

fn parse_n_trees_l(&mut self) -> Result<State, DecompressorError>

fn parse_n_trees_d(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_code_kind(&mut self) -> Result<PrefixCodeKind, DecompressorError>

fn parse_simple_prefix_code(&mut self, alphabet_size: usize) -> Result<HuffmanCodes, DecompressorError>

fn parse_complex_prefix_code(&mut self, h_skip: u8, alphabet_size: usize) -> Result<HuffmanCodes, DecompressorError>

fn parse_prefix_code(&mut self, alphabet_size: usize) -> Result<HuffmanCodes, DecompressorError>

fn parse_prefix_code_block_types_literals(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_code_block_counts_literals(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_code_block_types_insert_and_copy_lengths(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_code_block_counts_insert_and_copy_lengths(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_code_block_types_distances(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_code_block_counts_distances(&mut self) -> Result<State, DecompressorError>

fn parse_block_count(&mut self, prefix_code: &HuffmanCodes) -> Result<BLen, DecompressorError>

fn parse_first_block_count_literals(&mut self) -> Result<State, DecompressorError>

fn parse_first_block_count_insert_and_copy_lengths(&mut self) -> Result<State, DecompressorError>

fn parse_first_block_count_distances(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_codes_literals(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_codes_insert_and_copy_lengths(&mut self) -> Result<State, DecompressorError>

fn parse_prefix_codes_distances(&mut self) -> Result<State, DecompressorError>

fn parse_context_map(&mut self, n_trees: NTrees, len: usize) -> Result<ContextMap, DecompressorError>

fn parse_context_map_literals(&mut self) -> Result<State, DecompressorError>

fn parse_context_map_distances(&mut self) -> Result<State, DecompressorError>

fn inverse_move_to_front_transform(v: &mut [u8])

fn parse_insert_and_copy_length(&mut self) -> Result<State, DecompressorError>

fn decode_insert_and_copy_length(&mut self) -> Result<State, DecompressorError>

fn parse_block_switch_command(&mut self, prefix_tree_types: HuffmanCodes, btype: NBltypes, btype_prev: NBltypes, n_bltypes: NBltypes, prefix_tree_counts: HuffmanCodes) -> Result<BlockSwitch, DecompressorError>

fn parse_block_switch_command_literals(&mut self) -> Result<BlockSwitch, DecompressorError>

fn parse_block_switch_command_insert_and_copy_lengths(&mut self) -> Result<BlockSwitch, DecompressorError>

fn parse_block_switch_command_distances(&mut self) -> Result<BlockSwitch, DecompressorError>

fn parse_insert_literals(&mut self) -> Result<State, DecompressorError>

fn parse_distance_code(&mut self) -> Result<State, DecompressorError>

fn decode_distance(&mut self) -> Result<State, DecompressorError>

fn copy_literals(&mut self) -> Result<State, DecompressorError>

fn decompress(&mut self, buf: &mut [u8]) -> Result<usize, DecompressorError>