Module brotli::enc::combined_alloc

Macros

implement_allocator 🔒

Structs

CombiningAllocator

Traits

Allocator

BrotliAlloc

Functions

alloc_default 🔒

Helper function to create an empty allocator object

alloc_if 🔒

Helper function to allocate memory or return a default value if the condition is false

alloc_or_default 🔒

Helper function to allocate memory or return a default value when the size is 0.

allocate 🔒

Helper function to allocate memory with the given allocator (may crash if the len is 0).